requirements.txt
streamlit
pandas
matplotlib
Pillow
scikit-learn